Refactor and rename to dynamic deferred call
